Move Meeting Minutes section from HomePage to AboutPage
- Moved complete Meeting Minutes functionality to AboutPage (top of page)
- Removed Meeting Minutes section and unused imports from HomePage
- Updated NavigationButtons: Changed Memories Chinese label to '活动回忆'
- Made boto3 optional in server to support local development without AWS dependencies
