+++ ProjectCode/client/src/pages/HomePage.js
@@ -1,8 +1,4 @@
 import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
-  Alert,
   Box,
   Button,
   CircularProgress,
@@ -10,7 +6,6 @@ import {
   Dialog,
   DialogActions,
   DialogContent,
-  DialogContentText,
   DialogTitle,
   FormControl,
   IconButton,
@@ -24,25 +19,8 @@ import {
 import ArrowForwardIcon from '@mui/icons-material/ArrowForward';
 import EditIcon from '@mui/icons-material/Edit';
 import DragIndicatorIcon from '@mui/icons-material/DragIndicator';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import AddIcon from '@mui/icons-material/Add';
-import DeleteIcon from '@mui/icons-material/Delete';
-import SaveIcon from '@mui/icons-material/Save';
-import CancelIcon from '@mui/icons-material/Cancel';
-import { useEffect, useState, useCallback } from 'react';
+import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { getMeetingContent, getMeetingFiles } from '../api/meetings';
-import {
-  getAllMeetingMinutes,
-  createMeetingMinutes,
-  updateMeetingMinutes,
-  deleteMeetingMinutes
-} from '../api/meetingMinutes';
-import { safeMarkdown } from '../utils/markdown';
-import DOMPurify from 'dompurify';
-import ReactQuill from 'react-quill';
-import 'react-quill/dist/quill.snow.css';
-import { useAutoFillOnTab } from '../hooks';
 import {
   DndContext,
   closestCenter,
@@ -65,24 +43,6 @@ import { useAuth } from '../context/AuthContext';
 import { useAdmin } from '../context/AdminContext';
 import { getCarouselBanners } from '../api/banners';
 import { getActiveSections, updateSection, reorderSections, uploadImage } from '../api/homepageSections';
-
-// Quill editor configuration for meeting minutes
-const quillModules = {
-  toolbar: [
-    [{ 'header': [1, 2, 3, false] }],
-    ['bold', 'italic', 'underline'],
-    [{ 'list': 'ordered' }, { 'list': 'bullet' }],
-    [{ 'indent': '-1' }, { 'indent': '+1' }],
-    ['clean']
-  ],
-};
-
-const quillFormats = [
-  'header',
-  'bold', 'italic', 'underline',
-  'list', 'bullet',
-  'indent'
-];
 
 // Fallback carousel images if API fails
 const fallbackCarouselImages = [
@@ -330,226 +290,6 @@ export default function HomePage() {
   const [uploading, setUploading] = useState(false);
   const navigate = useNavigate();
 
-  // Meeting minutes state
-  const [meetings, setMeetings] = useState([]);
-  const [meetingsLoading, setMeetingsLoading] = useState(true);
-  const [meetingsError, setMeetingsError] = useState('');
-  const [meetingsSuccess, setMeetingsSuccess] = useState('');
-
-  // Meeting minutes editor state
-  const [isMeetingEditing, setIsMeetingEditing] = useState(false);
-  const [editingMeetingId, setEditingMeetingId] = useState(null);
-  const [meetingFormData, setMeetingFormData] = useState({
-    title: '',
-    meeting_date: new Date().toISOString().split('T')[0],
-    content: ''
-  });
-  const [meetingSaving, setMeetingSaving] = useState(false);
-
-  // Delete confirmation dialog for meetings
-  const [deleteMeetingDialog, setDeleteMeetingDialog] = useState({ open: false, id: null, title: '' });
-  const [deletingMeeting, setDeletingMeeting] = useState(false);
-
-  // Default values for Tab auto-fill
-  const meetingDefaultValues = {
-    title: 'Committee Meeting - [Month Year]'
-  };
-
-  const handleMeetingAutoFill = useAutoFillOnTab({
-    setValue: (field, value) => setMeetingFormData(prev => ({ ...prev, [field]: value })),
-    defaultValues: meetingDefaultValues
-  });
-
-  // Sanitize HTML content for display
-  const sanitizeMeetingHtml = (html) => {
-    return DOMPurify.sanitize(html, {
-      ALLOWED_TAGS: ['h1', 'h2', 'h3', 'p', 'strong', 'em', 'u', 'ol', 'ul', 'li', 'br'],
-      ALLOWED_ATTR: []
-    });
-  };
-
-  // Format date for display
-  const formatMeetingDate = (dateStr) => {
-    const date = new Date(dateStr + 'T00:00:00');
-    return date.toLocaleDateString('en-US', {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric'
-    });
-  };
-
-  // Fetch meeting minutes from database API
-  const fetchMeetingMinutes = useCallback(async () => {
-    try {
-      setMeetingsLoading(true);
-      const dbMeetings = await getAllMeetingMinutes();
-
-      // Also fetch local markdown files as fallback
-      const mdFiles = getMeetingFiles();
-      const localMeetings = await Promise.all(
-        mdFiles.map(async (filename) => {
-          try {
-            const content = await getMeetingContent(filename);
-            const title = content.split('\n')[0].replace('# ', '');
-            const date = filename.split('.')[0];
-            return {
-              title,
-              content,
-              meeting_date: date,
-              filename,
-              isLocal: true
-            };
-          } catch (error) {
-            return null;
-          }
-        })
-      );
-
-      // Combine database meetings with local ones
-      const validLocalMeetings = localMeetings.filter(m => m !== null);
-
-      // Database meetings take priority - format them consistently
-      const formattedDbMeetings = dbMeetings.map(m => ({
-        ...m,
-        isLocal: false
-      }));
-
-      // Combine and sort by date (most recent first)
-      const allMeetings = [...formattedDbMeetings, ...validLocalMeetings];
-      allMeetings.sort((a, b) => new Date(b.meeting_date) - new Date(a.meeting_date));
-
-      setMeetings(allMeetings);
-      setMeetingsError('');
-    } catch (err) {
-      console.error('Error fetching meeting minutes:', err);
-      // Fall back to local files only
-      try {
-        const mdFiles = getMeetingFiles();
-        const localMeetings = await Promise.all(
-          mdFiles.map(async (filename) => {
-            try {
-              const content = await getMeetingContent(filename);
-              const title = content.split('\n')[0].replace('# ', '');
-              const date = filename.split('.')[0];
-              return { title, content, meeting_date: date, filename, isLocal: true };
-            } catch (error) {
-              return null;
-            }
-          })
-        );
-        const validMeetings = localMeetings.filter(m => m !== null);
-        validMeetings.sort((a, b) => new Date(b.meeting_date) - new Date(a.meeting_date));
-        setMeetings(validMeetings);
-      } catch (localErr) {
-        console.error('Error fetching local meetings:', localErr);
-      }
-    } finally {
-      setMeetingsLoading(false);
-    }
-  }, []);
-
-  // Meeting minutes editor handlers
-  const handleStartNewMeeting = () => {
-    setIsMeetingEditing(true);
-    setEditingMeetingId(null);
-    setMeetingFormData({
-      title: '',
-      meeting_date: new Date().toISOString().split('T')[0],
-      content: ''
-    });
-  };
-
-  const handleStartEditMeeting = (meeting) => {
-    if (meeting.isLocal) {
-      setMeetingsError('Local files cannot be edited. Only database entries can be modified.');
-      setTimeout(() => setMeetingsError(''), 5000);
-      return;
-    }
-    setIsMeetingEditing(true);
-    setEditingMeetingId(meeting.id);
-    setMeetingFormData({
-      title: meeting.title,
-      meeting_date: meeting.meeting_date,
-      content: meeting.content
-    });
-  };
-
-  const handleCancelMeeting = () => {
-    setIsMeetingEditing(false);
-    setEditingMeetingId(null);
-    setMeetingFormData({
-      title: '',
-      meeting_date: new Date().toISOString().split('T')[0],
-      content: ''
-    });
-  };
-
-  const handleSaveMeeting = async () => {
-    if (!meetingFormData.title.trim()) {
-      setMeetingsError('Please enter a title. / 请输入标题。');
-      return;
-    }
-    if (!meetingFormData.meeting_date) {
-      setMeetingsError('Please select a meeting date. / 请选择会议日期。');
-      return;
-    }
-    if (!meetingFormData.content.trim() || meetingFormData.content === '<p><br></p>') {
-      setMeetingsError('Please enter meeting minutes content. / 请输入会议纪要内容。');
-      return;
-    }
-
-    setMeetingSaving(true);
-    setMeetingsError('');
-
-    try {
-      if (editingMeetingId) {
-        await updateMeetingMinutes(editingMeetingId, meetingFormData, currentUser.uid);
-        setMeetingsSuccess('Meeting minutes updated successfully! / 会议纪要更新成功！');
-      } else {
-        await createMeetingMinutes(meetingFormData, currentUser.uid);
-        setMeetingsSuccess('Meeting minutes created successfully! / 会议纪要创建成功！');
-      }
-
-      await fetchMeetingMinutes();
-      handleCancelMeeting();
-      setTimeout(() => setMeetingsSuccess(''), 5000);
-    } catch (err) {
-      console.error('Error saving meeting minutes:', err);
-      setMeetingsError('Failed to save meeting minutes. Please try again. / 保存失败，请重试。');
-    } finally {
-      setMeetingSaving(false);
-    }
-  };
-
-  const handleDeleteMeetingClick = (meeting) => {
-    if (meeting.isLocal) {
-      setMeetingsError('Local files cannot be deleted through the app.');
-      setTimeout(() => setMeetingsError(''), 5000);
-      return;
-    }
-    setDeleteMeetingDialog({ open: true, id: meeting.id, title: meeting.title });
-  };
-
-  const handleDeleteMeetingConfirm = async () => {
-    setDeletingMeeting(true);
-    try {
-      await deleteMeetingMinutes(deleteMeetingDialog.id, currentUser.uid);
-      setMeetingsSuccess('Meeting minutes deleted successfully! / 会议纪要已删除！');
-      await fetchMeetingMinutes();
-      setTimeout(() => setMeetingsSuccess(''), 5000);
-    } catch (err) {
-      console.error('Error deleting meeting minutes:', err);
-      setMeetingsError('Failed to delete meeting minutes. / 删除失败。');
-    } finally {
-      setDeletingMeeting(false);
-      setDeleteMeetingDialog({ open: false, id: null, title: '' });
-    }
-  };
-
-  const handleDeleteMeetingCancel = () => {
-    setDeleteMeetingDialog({ open: false, id: null, title: '' });
-  };
-
   // Drag and drop sensors
   const sensors = useSensors(
     useSensor(PointerSensor),
@@ -580,11 +320,6 @@ export default function HomePage() {
 
     fetchData();
   }, []);
-
-  // Fetch meeting minutes
-  useEffect(() => {
-    fetchMeetingMinutes();
-  }, [fetchMeetingMinutes]);
 
   // Auto-rotate carousel
   useEffect(() => {
@@ -926,288 +661,6 @@ export default function HomePage() {
           </Box>
         </Container>
       )}
-
-      {/* Meeting Minutes Section */}
-      <Container maxWidth="xl" sx={{ px: { xs: 1, sm: 2 }, mt: { xs: 3, sm: 4 } }}>
-        <Typography
-          variant="h6"
-          sx={{
-            fontWeight: 600,
-            color: '#FFA500',
-            mb: 2,
-            fontSize: { xs: '1rem', sm: '1.25rem' }
-          }}
-        >
-          Meeting Minutes 会议纪要
-        </Typography>
-
-        {adminModeEnabled && !isMeetingEditing && (
-          <Box sx={{ display: 'flex', justifyContent: 'flex-end', mb: { xs: 2, sm: 3 } }}>
-            <Button
-              variant="contained"
-              startIcon={<AddIcon />}
-              onClick={handleStartNewMeeting}
-              sx={{
-                backgroundColor: '#FFB84D',
-                color: 'white',
-                textTransform: 'none',
-                '&:hover': {
-                  backgroundColor: '#FFA833',
-                }
-              }}
-            >
-              Add Meeting Minutes
-            </Button>
-          </Box>
-        )}
-
-        {!adminModeEnabled && <Box sx={{ mb: { xs: 2, sm: 3 } }} />}
-
-        {/* Success/Error Messages */}
-        {meetingsSuccess && (
-          <Alert severity="success" sx={{ mb: 2 }} onClose={() => setMeetingsSuccess('')}>
-            {meetingsSuccess}
-          </Alert>
-        )}
-        {meetingsError && (
-          <Alert severity="error" sx={{ mb: 2 }} onClose={() => setMeetingsError('')}>
-            {meetingsError}
-          </Alert>
-        )}
-
-        {/* Editor Form */}
-        {isMeetingEditing && (
-          <Box sx={{
-            mb: 3,
-            p: { xs: 2, sm: 3 },
-            backgroundColor: 'white',
-            borderRadius: '12px',
-            boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)',
-            border: '2px solid #FFA500'
-          }}>
-            <Typography variant="h6" sx={{ mb: 2, fontWeight: 600, color: '#333' }}>
-              {editingMeetingId ? 'Edit Meeting Minutes / 编辑会议纪要' : 'New Meeting Minutes / 新会议纪要'}
-            </Typography>
-
-            <TextField
-              name="title"
-              fullWidth
-              label="Title / 标题"
-              value={meetingFormData.title}
-              onChange={(e) => setMeetingFormData({ ...meetingFormData, title: e.target.value })}
-              onKeyDown={handleMeetingAutoFill}
-              sx={{ mb: 2 }}
-              placeholder={meetingDefaultValues.title}
-            />
-
-            <TextField
-              fullWidth
-              type="date"
-              label="Meeting Date / 会议日期"
-              value={meetingFormData.meeting_date}
-              onChange={(e) => setMeetingFormData({ ...meetingFormData, meeting_date: e.target.value })}
-              sx={{ mb: 2 }}
-              InputLabelProps={{ shrink: true }}
-            />
-
-            <Typography variant="body2" color="text.secondary" sx={{ mb: 1 }}>
-              Content / 内容 (supports bold, lists, and formatting)
-            </Typography>
-            <Box sx={{ mb: 2, '.ql-container': { minHeight: '200px' }, backgroundColor: 'white' }}>
-              <ReactQuill
-                theme="snow"
-                value={meetingFormData.content}
-                onChange={(content) => setMeetingFormData({ ...meetingFormData, content })}
-                modules={quillModules}
-                formats={quillFormats}
-                placeholder="Enter meeting minutes here... / 在此输入会议纪要..."
-              />
-            </Box>
-
-            <Box sx={{ display: 'flex', gap: 2, justifyContent: 'flex-end' }}>
-              <Button
-                startIcon={<CancelIcon />}
-                onClick={handleCancelMeeting}
-                disabled={meetingSaving}
-              >
-                Cancel / 取消
-              </Button>
-              <Button
-                variant="contained"
-                startIcon={meetingSaving ? <CircularProgress size={20} color="inherit" /> : <SaveIcon />}
-                onClick={handleSaveMeeting}
-                disabled={meetingSaving}
-                sx={{
-                  backgroundColor: '#FFA500',
-                  '&:hover': { backgroundColor: '#FF8C00' }
-                }}
-              >
-                {meetingSaving ? 'Saving...' : 'Save / 保存'}
-              </Button>
-            </Box>
-          </Box>
-        )}
-
-        {meetingsLoading ? (
-          <Box sx={{ display: 'flex', justifyContent: 'center', py: 4 }}>
-            <CircularProgress sx={{ color: '#FFA500' }} />
-          </Box>
-        ) : meetings.length === 0 ? (
-          <Typography variant="body1" color="text.secondary" sx={{ textAlign: 'center' }}>
-            No meeting minutes available.
-          </Typography>
-        ) : (
-          <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-            {meetings.map((meeting) => (
-              <Accordion
-                key={meeting.id || meeting.filename}
-                defaultExpanded={false}
-                sx={{
-                  backgroundColor: 'white',
-                  borderRadius: '12px !important',
-                  boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)',
-                  '&:before': {
-                    display: 'none',
-                  },
-                  '&.Mui-expanded': {
-                    margin: '0',
-                  }
-                }}
-              >
-                <AccordionSummary
-                  expandIcon={<ExpandMoreIcon />}
-                  sx={{
-                    '& .MuiAccordionSummary-content': {
-                      margin: '12px 0',
-                      display: 'flex',
-                      justifyContent: 'space-between',
-                      alignItems: 'center',
-                      pr: 1
-                    },
-                    '& .MuiAccordionSummary-expandIconWrapper': {
-                      color: '#FFA500',
-                    }
-                  }}
-                >
-                  <Box sx={{ flex: 1 }}>
-                    <Typography
-                      variant="h6"
-                      sx={{
-                        fontWeight: 600,
-                        color: '#333',
-                        fontSize: { xs: '1rem', sm: '1.25rem' }
-                      }}
-                    >
-                      {meeting.title}
-                    </Typography>
-                    <Typography variant="body2" color="text.secondary">
-                      {formatMeetingDate(meeting.meeting_date)}
-                      {meeting.isLocal && ' (Local file)'}
-                    </Typography>
-                  </Box>
-
-                  {adminModeEnabled && !meeting.isLocal && (
-                    <Box sx={{ display: 'flex', gap: 0.5, ml: 2 }} onClick={(e) => e.stopPropagation()}>
-                      <Tooltip title="Edit / 编辑">
-                        <IconButton
-                          size="small"
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            handleStartEditMeeting(meeting);
-                          }}
-                          disabled={isMeetingEditing}
-                          sx={{ color: 'primary.main' }}
-                        >
-                          <EditIcon fontSize="small" />
-                        </IconButton>
-                      </Tooltip>
-                      <Tooltip title="Delete / 删除">
-                        <IconButton
-                          size="small"
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            handleDeleteMeetingClick(meeting);
-                          }}
-                          disabled={isMeetingEditing}
-                          sx={{ color: 'error.main' }}
-                        >
-                          <DeleteIcon fontSize="small" />
-                        </IconButton>
-                      </Tooltip>
-                    </Box>
-                  )}
-                </AccordionSummary>
-                <AccordionDetails>
-                  <Box
-                    dangerouslySetInnerHTML={{
-                      __html: meeting.isLocal ? safeMarkdown(meeting.content) : sanitizeMeetingHtml(meeting.content)
-                    }}
-                    sx={{
-                      '& h1': {
-                        fontSize: '1.8rem',
-                        fontWeight: 600,
-                        color: '#333',
-                        marginBottom: '1rem'
-                      },
-                      '& h2': {
-                        fontSize: '1.5rem',
-                        fontWeight: 600,
-                        color: '#444',
-                        marginBottom: '0.8rem'
-                      },
-                      '& h3': {
-                        fontSize: '1.2rem',
-                        fontWeight: 600,
-                        color: '#555',
-                        marginBottom: '0.6rem'
-                      },
-                      '& p': {
-                        marginBottom: '0.5rem'
-                      },
-                      '& ul, & ol': {
-                        paddingLeft: '1.5rem',
-                        marginBottom: '1rem'
-                      },
-                      '& li': {
-                        marginBottom: '0.5rem'
-                      }
-                    }}
-                  />
-                </AccordionDetails>
-              </Accordion>
-            ))}
-          </Box>
-        )}
-      </Container>
-
-      {/* Delete Meeting Confirmation Dialog */}
-      <Dialog open={deleteMeetingDialog.open} onClose={handleDeleteMeetingCancel}>
-        <DialogTitle>
-          Delete Meeting Minutes?
-          <br />
-          删除会议纪要？
-        </DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            Are you sure you want to delete "{deleteMeetingDialog.title}"? This action cannot be undone.
-            <br /><br />
-            您确定要删除"{deleteMeetingDialog.title}"吗？此操作无法撤销。
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleDeleteMeetingCancel} disabled={deletingMeeting}>
-            Cancel / 取消
-          </Button>
-          <Button
-            onClick={handleDeleteMeetingConfirm}
-            variant="contained"
-            color="error"
-            disabled={deletingMeeting}
-          >
-            {deletingMeeting ? <CircularProgress size={20} /> : 'Delete / 删除'}
-          </Button>
-        </DialogActions>
-      </Dialog>
 
       {/* Dynamic Homepage Sections with Drag-and-Drop for Admins */}
       {adminModeEnabled ? (
